Move secondary enhanced components into own file; refine date as prop usage
