+++ src/applications/user-testing/disability-benefits-all-claims-conditions-chapter/pages/shared/index.js
@@ -68,6 +68,7 @@ const clearSideOfBody = (formData, index, setFormData) => {
   });
 };
 
+// Just for SecondaryEnhanced demo
 const clearConditionNotListed = (formData, setFormData) => {
   setFormData({
     ...formData,
